+++ tests/test_analysis/test_forecast/test_plots.py
@@ -1,6 +1,7 @@
 import pandas as pd
 import pytest
 
+from etna.analysis import metric_per_segment_distribution_plot
 from etna.analysis import plot_metric_per_segment
 from etna.analysis import plot_residuals
 from etna.analysis.forecast.plots import _get_borders_comparator
@@ -80,7 +81,7 @@ def metrics_df_no_folds(metrics_df_with_folds) -> pd.DataFrame:
     "df_name, metric_name",
     [
         ("metrics_df_with_folds", "MAE"),
-        ("metrics_df_no_folds", "MSE"),
+        ("metrics_df_no_folds", "MAE"),
         ("metrics_df_no_folds", "MSE"),
     ],
 )
@@ -112,3 +113,49 @@ def test_plot_metric_per_segment_warning_non_comparable_segments(df_name, metric
     metrics_df = request.getfixturevalue(df_name)
     with pytest.warns(UserWarning, match="Some segments have different set of folds to be aggregated on"):
         plot_metric_per_segment(metrics_df=metrics_df, metric_name=metric_name)
+
+
+@pytest.mark.parametrize("plot_type", ["hist", "box", "violin"])
+@pytest.mark.parametrize(
+    "df_name, metric_name, per_fold_aggregation_mode",
+    [
+        ("metrics_df_with_folds", "MAE", None),
+        ("metrics_df_with_folds", "MAE", "mean"),
+        ("metrics_df_with_folds", "MAE", "median"),
+        ("metrics_df_no_folds", "MAE", None),
+        ("metrics_df_no_folds", "MSE", None),
+    ],
+)
+def test_plot_metric_per_segment_ok(df_name, metric_name, per_fold_aggregation_mode, plot_type, request):
+    metrics_df = request.getfixturevalue(df_name)
+    metric_per_segment_distribution_plot(
+        metrics_df=metrics_df,
+        metric_name=metric_name,
+        per_fold_aggregation_mode=per_fold_aggregation_mode,
+        plot_type=plot_type,
+    )
+
+
+@pytest.mark.parametrize(
+    "df_name, metric_name",
+    [
+        ("metrics_df_with_folds", "MAPE"),
+        ("metrics_df_no_folds", "RMSE"),
+    ],
+)
+def test_plot_metric_per_segment_warning_empty_segments(df_name, metric_name, request):
+    metrics_df = request.getfixturevalue(df_name)
+    with pytest.warns(UserWarning, match="There are segments with all missing metric values"):
+        metric_per_segment_distribution_plot(metrics_df=metrics_df, metric_name=metric_name)
+
+
+@pytest.mark.parametrize(
+    "df_name, metric_name",
+    [
+        ("metrics_df_with_folds", "MSE"),
+    ],
+)
+def test_plot_metric_per_segment_warning_non_comparable_segments(df_name, metric_name, request):
+    metrics_df = request.getfixturevalue(df_name)
+    with pytest.warns(UserWarning, match="Some segments have different set of folds to be aggregated on"):
+        metric_per_segment_distribution_plot(metrics_df=metrics_df, metric_name=metric_name)